feat: implement modules subtask
