+++ src/modules/esm.mjs
@@ -1,0 +1,43 @@
+import { sep, dirname } from 'path';
+import { release, version } from 'os';
+import { fileURLToPath } from 'url';
+import { createServer as createServerHttp } from 'http';
+import { createRequire } from 'module';
+import './files/c.js';
+
+const random = Math.random();
+
+let unknownObject;
+
+const moduleRequire = createRequire(import.meta.url);
+
+if (random > 0.5) {
+  unknownObject = moduleRequire('./files/a.json');
+} else {
+  unknownObject = moduleRequire('./files/b.json');
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${sep}"`);
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { unknownObject, myServer };
